Fixes #1297

I will test the performance using the test code from #1310