@tangruize 2024-09-10
The current condition for checking if a read is lockable is `lock & (WRITER | MAX_READER | BEING_UPGRADED) == 0`, which does not fully utilize the 61 bits for the maximum reader lock count. This might have been considered in the previous `try_read()` method, where concurrent momentarily increasing the count could overflow if we compared something like `(lock & MASK) < MAX_READERS`.

This fix addressed it, therefore we can define `const MAX_READERS: usize = (1 << (usize::BITS - 3)) - 1`, and then use the condition `(lock ^ MAX_READERS != 0) && (lock & (WRITER | BEING_UPGRADED) == 0)` for `try_read()`.