@tangruize 2024-09-11
Here is the performance test result:

### After fix

`RwLock::read`
time consuming: `[152ns, 151ns, 151ns]`

`RwLock::read_irq_disabled`
time consuming: `[124ns, 119ns, 118ns]`

### Before fix

`RwLock::read_before_fix`
time consuming: `[112ns, 109ns, 111ns]`

`RwLock::read_irq_disabled_before_fix`
time consuming: `[83ns, 82ns, 82ns]`