@tangruize 2024-09-14
To make it more efficient, we can "inline" the `fetch_update()` method, which eliminate two closures. Below are the performance test results ([this branch](https://github.com/tangruize/asterinas/tree/fix_rwlock_effi)):

After fix:
`RwLock::read`                                                                                                                                                                                                  
time consuming: `[136ns, 135ns, 132ns]`                                                                                                                                                                         

`RwLock::read_irq_disabled`                                                                                                                                                                                     
time consuming: `[106ns, 105ns, 105ns]`                                                                                                                                                                         

Before fix:
`RwLock::read_before_fix`                                                                                                                                                                                       
time consuming: `[117ns, 117ns, 116ns]`                                                                                                                                                                         

`RwLock::read_irq_disabled_before_fix`
time consuming: `[92ns, 88ns, 91ns]`