@tangruize 2024-09-14
> Do you think that you can make RwLock correct but also as efficient as it is used to be?

@tatetian It doesn't seem possible to be more efficient using the `load` and `compare_exchange` methods. I experimented with removing the loop and condition checks in the `try_read()` method (note: it was an incorrect implementation), and the performance remained almost the same. This shows that the two atomic operations (`load` and `compare_exchange`) have a constant overhead compared to using only one (`fetch_add`).

I referenced similar implementations in [Rust STD](https://stdrs.dev/nightly/x86_64-unknown-linux-gnu/src/std/sys/unix/locks/futex_rwlock.rs.html#70-88) and [Linux](https://github.com/torvalds/linux/blob/b7718454f937f50f44f98c1222f5135eaef29132/kernel/locking/rwbase_rt.c#L53-L66). They also use `load` and `compare_exchange`. They design fast paths to optimize performance.