@tatetian 2024-11-06
> To make it more efficient, we can "inline" the `fetch_update()` method, which eliminate two closures. Below are the performance test results ([tangruize@faf11bb](https://github.com/tangruize/asterinas/commit/faf11bbaa7e4503278bfad510fca785f0d2826b2)):
> 
> After fix: `RwLock::read` time consuming: `[136ns, 135ns, 132ns]`
> 
> `RwLock::read_irq_disabled` time consuming: `[106ns, 105ns, 105ns]`
> 
> Before fix: `RwLock::read_before_fix` time consuming: `[117ns, 117ns, 116ns]`
> 
> `RwLock::read_irq_disabled_before_fix` time consuming: `[92ns, 88ns, 91ns]`

My major concern with the changes is that it degrades the performance of the common path (only readers exist, the best case for read-write locks) to improve that of the rare path (no one is using downgrade at the moment).

Do you have any results compare to the read-write lock in Linux? It looks that our new implementation adopts the same logic (read and then compare-exchange). If our new version of read-write locks perform as well as Linux, then I would be happy to accept it. If not, then I will be hesitate to do so.

If you can't come up with a more efficient implementation that is on par with Linux in the near future, then I think a compromised choice would be accept the changes proposed by this PR as a compile-time, non-default feature and add a TODO to emphasize that the performance pros and cons so that we can fix it in the future.