@tangruize 2024-11-07
Yes, I also think this is a trade-off between performance and absolute correctness.

For now, we can keep the current implementation because Asterinas doesn't use the `downgrade()` method. However, other users may use ostd differently or require `downgrade()`. For those users, we should document potential contention scenarios.

In the longer term, we should move towards RCU to replace rwlock, as it offers significant performance and scalability improvements in read-heavy workloads.

Supporting Details:

Linux's rwlock implementation does not support `upgrade()` and `downgrade()` operations. The `queued_read_lock()` and `queued_read_trylock()` functions use `atomic_add()` and `atomic_sub()` operations (with an extra `atomic_read()` for `queued_read_trylock()` for checking purposes, <https://elixir.bootlin.com/linux/v6.11.3/source/include/asm-generic/qrwlock.h#L82>):

```c
/**
 * queued_read_lock - acquire read lock of a queued rwlock
 * @lock: Pointer to queued rwlock structure
 */
static inline void queued_read_lock(struct qrwlock *lock)
{
	int cnts;

	cnts = atomic_add_return_acquire(_QR_BIAS, &lock->cnts);
	if (likely(!(cnts & _QW_WMASK)))
		return;

	/* The slowpath will decrement the reader count, if necessary. */
	queued_read_lock_slowpath(lock);
}
/**
 * queued_read_trylock - try to acquire read lock of a queued rwlock
 * @lock : Pointer to queued rwlock structure
 * Return: 1 if lock acquired, 0 if failed
 */
static inline int queued_read_trylock(struct qrwlock *lock)
{
	int cnts;

	cnts = atomic_read(&lock->cnts);
	if (likely(!(cnts & _QW_WMASK))) {
		cnts = (u32)atomic_add_return_acquire(_QR_BIAS, &lock->cnts);
		if (likely(!(cnts & _QW_WMASK)))
			return 1;
		atomic_sub(_QR_BIAS, &lock->cnts);
	}
	return 0;
}
```

In the Linux documentation (<https://docs.kernel.org/locking/spinlocks.html#lesson-2-reader-writer-spinlocks>) (maybe `trylock()` add an additional `atomic_read()`):

> Reader-writer locks require more atomic memory operations than simple spinlocks. Unless the reader critical section is long, you are better off just using spinlocks.

Moreover, Linux is replacing rwlock with RCU:

> NOTE! We are working hard to remove reader-writer spinlocks in most cases, so please don't add a new one without consensus. (Instead, see RCU Concepts for complete information.)