@tatetian 2024-11-08
I think you have brought up two good points.

> Linux's rwlock implementation does not support upgrade() and downgrade() operations.

> Moreover, Linux is replacing rwlock with RCU

Maybe we don't need read-write spinlocks or make them upgradable in the first place.